Move reader macros
