+++ src/parser/reader-macros/object-literal.ts
