+++ src/parser/reader-macros/types.ts
@@ -1,6 +1,6 @@
-import { File } from "../lib/file.js";
-import { Expr, List } from "../syntax-objects/index.js";
-import { Token } from "../lib/token.js";
+import { File } from "../../lib/file.js";
+import { Expr, List } from "../../syntax-objects/index.js";
+import { Token } from "../../lib/token.js";
 
 export interface ReaderMacro {
   match: (token: Token, prev?: Token) => boolean;
